+++ apps/official-site-e2e/src/integration/app.spec.ts
@@ -8,6 +8,6 @@ describe('official-site', () => {
     cy.login('my-email@something.com', 'myPassword');
 
     // Function helper example, see `../support/app.po.ts` file
-    getGreeting().contains('Welcome to official-site!');
+    getGreeting().contains('Welcome official-site');
   });
 });