+++ libs/angular-zen/core/src/rxjs/observe/directives/observe-concat.directive.spec.ts
@@ -1,0 +1,8 @@
+// import { ObserveLatestDirective } from './observe-latest.directive';
+
+// describe('ObserveLatestDirective', () => {
+//   it('should create an instance', () => {
+//     const directive = new ObserveLatestDirective();
+//     expect(directive).toBeTruthy();
+//   });
+// });